+++ Database/TableInterface.php
@@ -18,7 +18,7 @@ defined('JPATH_PLATFORM') or die;
  * @package         Prism
  * @subpackage      Database\Interfaces
  *
- * @deprecated since v1.15
+ * @deprecated v1.15
  */
 interface TableInterface
 {