+++ ui/helpers/ui.php
@@ -26,12 +26,58 @@ abstract class PrismUI
     protected static $loaded = array();
 
     /**
+     * Include backend styles.
+     *
+     * <code>
+     * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
+     *
+     * JHtml::_('Prism.ui.backendStyles');
+     * </code>
+     */
+    public static function backendStyles()
+    {
+        // Only load once
+        if (!empty(self::$loaded[__METHOD__])) {
+            return;
+        }
+
+        $document = JFactory::getDocument();
+
+        $document->addStyleSheet(JUri::root() . 'libraries/Prism/ui/css/backend.style.css');
+
+        self::$loaded[__METHOD__] = true;
+    }
+
+    /**
+     * Include some general styles.
+     *
+     * <code>
+     * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
+     *
+     * JHtml::_('Prism.ui.styles');
+     * </code>
+     */
+    public static function styles()
+    {
+        // Only load once
+        if (!empty(self::$loaded[__METHOD__])) {
+            return;
+        }
+
+        $document = JFactory::getDocument();
+
+        $document->addStyleSheet(JUri::root() . 'libraries/Prism/ui/css/styles.css');
+
+        self::$loaded[__METHOD__] = true;
+    }
+    
+    /**
      * Include jQuery PNotify library.
      *
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.pnotify');
+     * JHtml::_('Prism.ui.pnotify');
      * </code>
      *
      * @link http://sciactive.github.io/pnotify/ Documentation of PNotify
@@ -57,7 +103,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap2Editable');
+     * JHtml::_('Prism.ui.bootstrap2Editable');
      * </code>
      *
      * @link https://github.com/vitalets/x-editable Documentation of Bootstrap Editable
@@ -84,7 +130,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap3Editable');
+     * JHtml::_('Prism.ui.bootstrap3Editable');
      * </code>
      *
      * @link https://github.com/vitalets/x-editable Documentation of Bootstrap Editable
@@ -111,7 +157,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrapMaxLength');
+     * JHtml::_('Prism.ui.bootstrapMaxLength');
      * </code>
      *
      * @link https://github.com/mimo84/bootstrap-maxlength Documentation of Bootstrap Maxlength
@@ -136,7 +182,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap2FileInput');
+     * JHtml::_('Prism.ui.bootstrap2FileInput');
      * </code>
      */
     public static function bootstrap2FileInput()
@@ -160,7 +206,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap3FileInput');
+     * JHtml::_('Prism.ui.bootstrap3FileInput');
      * </code>
      *
      * @link https://github.com/kartik-v/bootstrap-fileinput Documentation of Bootstrap Fileinput
@@ -187,7 +233,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrapDatepicker');
+     * JHtml::_('Prism.ui.bootstrapDatepicker');
      * </code>
      *
      * @link http://eonasdan.github.io/bootstrap-datetimepicker/ Documentation of Bootstrap Datepicker
@@ -215,7 +261,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap2Typeahead');
+     * JHtml::_('Prism.ui.bootstrap2Typeahead');
      * </code>
      *
      * @link http://plugins.upbootstrap.com/bootstrap-ajax-typeahead/#docs Documentation of Bootstrap 2 Typeahead
@@ -239,7 +285,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.bootstrap3Typeahead');
+     * JHtml::_('Prism.ui.bootstrap3Typeahead');
      * </code>
      *
      * @link https://github.com/twitter/typeahead.js Documentation of Bootstrap 3 Typeahead
@@ -264,7 +310,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.parsley');
+     * JHtml::_('Prism.ui.parsley');
      * </code>
      *
      * @link http://parsleyjs.org/ Documentation of Parsley
@@ -290,7 +336,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.joomlaHelper');
+     * JHtml::_('Prism.ui.joomlaHelper');
      * </code>
      */
     public static function joomlaHelper()
@@ -312,7 +358,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.fileupload');
+     * JHtml::_('Prism.ui.fileupload');
      * </code>
      *
      * @link http://blueimp.github.io/jQuery-File-Upload/ Documentation of jQuery File Upload
@@ -341,7 +387,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.d3');
+     * JHtml::_('Prism.ui.d3');
      * </code>
      *
      * @param bool $cdn Include the library from content delivery network.
@@ -372,7 +418,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.cropper');
+     * JHtml::_('Prism.ui.cropper');
      * </code>
      *
      * @link http://fengyuanchen.github.io/cropper/ Cropper documentation
@@ -398,7 +444,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.joomlaList');
+     * JHtml::_('Prism.ui.joomlaList');
      * </code>
      */
     public static function joomlaList()
@@ -420,7 +466,7 @@ abstract class PrismUI
      * <code>
      * JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');
      *
-     * JHtml::_('prism.ui.serializeJson');
+     * JHtml::_('Prism.ui.serializeJson');
      * </code>
      */
     public static function serializeJson()
@@ -477,7 +523,7 @@ abstract class PrismUI
         }
 
         // Load the calendar behavior
-        JHtml::_('prism.ui.bootstrap3Datepicker');
+        JHtml::_('Prism.ui.bootstrap3Datepicker');
         $languageTag = JFactory::getLanguage()->getTag();
         $locale      = substr($languageTag, 0, 2);